+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 (Add entries below in the order of server, console, cli, docs, others)
 
 - console: add `pool_timeout`, `connection_lifetime` and `isolation_level` connection params to connect database form
+- console: add check constraints and comments to MS SQL Server tables' read-only modify page
 
 ## v2.0.0-alpha.11